Simplified code
Now it only supports 1 led. The code had to be simplified because in
the blinkLED function the stack size was overflowing. Too many nested
functions and variable allocations were happening and the code was
crashing sometimes. The stack size was too close to the maximum and the
effective stack usage was depending on the inlining of the compiler not
always happening.